template/pyproject.toml
[project]
name = "research-template"
version = "0.1.0"
description = "Educational research template for ML/DL learning module"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Asisten Dosen AI Module" }]
license = { text = "MIT" }

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "tensorboard>=2.14",
    "tqdm>=4.66",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "scikit-learn>=1.3",
    "pillow>=10.0",
]

[project.optional-dependencies]
demo = [
    "streamlit>=1.28",
    "gradio>=4.0",
]
medical = [
    "medmnist>=2.2",
]
dev = [
    "jupyter>=1.0",
    "ipykernel>=6.25",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"
